refactor graphql responses generation
